requirements.txt
faker
pandas
google-cloud-bigquery
pyarrow
